nearly last round of fixing compilation warnings

+++ src/apps/dirac/fcwf.cc
@@ -344,7 +344,7 @@ Fcwf copy(Fcwf psi){
 std::complex<double> inner(std::vector<Fcwf>& a, std::vector<Fcwf>& b){
      MADNESS_ASSERT(a.size() == b.size());
      std::complex<double> result(0,0);
-     for(int i = 0; i < a.size(); i++){
+     for(size_t i = 0; i < a.size(); i++){
           result += inner(a[i],b[i]);    
      }
      return result;
@@ -354,7 +354,7 @@ std::complex<double> inner(std::vector<Fcwf>& a, std::vector<Fcwf>& b){
 std::vector<Fcwf> operator*(const std::vector<Fcwf>& psis, std::complex<double> a){
      std::vector<Fcwf> result;
      if(psis.size() != 0){
-          for(int i = 0; i < psis.size(); i++){
+          for(size_t i = 0; i < psis.size(); i++){
                result.push_back(psis[i]*a);
           }
      }
@@ -364,7 +364,7 @@ std::vector<Fcwf> operator*(const std::vector<Fcwf>& psis, std::complex<double> 
 std::vector<Fcwf> operator*(std::complex<double> a, const std::vector<Fcwf>& psis){
      std::vector<Fcwf> result;
      if(psis.size() != 0){
-          for(int i = 0; i < psis.size(); i++){
+          for(size_t i = 0; i < psis.size(); i++){
                result.push_back(psis[i]*a);
           }
      }
@@ -378,7 +378,7 @@ void operator+=(std::vector<Fcwf>& phi, const std::vector<Fcwf>& psi){
      }
      else if(psi.size() != 0){
           MADNESS_ASSERT(phi.size()==psi.size());
-          for(int i=0; i < psi.size(); i++){
+          for(size_t i=0; i < psi.size(); i++){
                phi[i]+=psi[i];
           }
      }
@@ -395,7 +395,7 @@ std::vector<Fcwf> operator-(const std::vector<Fcwf>& phi, const std::vector<Fcwf
      }
      else{
           MADNESS_ASSERT(phi.size()==psi.size());
-          for(int i=0; i < psi.size(); i++){
+          for(size_t i=0; i < psi.size(); i++){
                result.push_back(phi[i]-psi[i]);
           }
      }
@@ -411,7 +411,7 @@ Fcwf_vector_allocator::Fcwf_vector_allocator(World& world, unsigned int m_size)
 //Overloading () operator
 std::vector<Fcwf> Fcwf_vector_allocator::operator()(){
      std::vector<Fcwf> result;
-     for(int i=0; i < m_size; i++){
+     for(unsigned int i=0; i < m_size; i++){
           result.push_back(Fcwf(world));
      }
      return result;
